+++ DiemDuLich/EntityModel/EFModel/Category.cs
@@ -28,7 +28,7 @@ namespace EntityModel.EFModel
 
         public DateTime? CreateDate { get; set; }
 
-        public int? UserID { get; set; }
+        public long? UserID { get; set; }
 
         public DateTime? ModifiedDate { get; set; }
